Version 2.8 - switch to the maxmod audio library!

+++ arm9/source/emu/pokeysnd.h
@@ -110,7 +110,7 @@ extern int stereo_enabled;
 extern int serio_sound_enabled;
 extern int console_sound_enabled;
 
-extern void (*Pokey_process_ptr)(void *sndbuffer, unsigned int sndn);
+extern void (*Pokey_process_ptr)(void *sndbuffer, uint16 sndn);
 extern void (*Update_pokey_sound)(uint16 addr, uint8 val, uint8 /*chip*/, uint8 gain);
 extern void (*Update_serio_sound)(int out, UBYTE data);
 extern void (*Update_consol_sound)(int set);
@@ -119,7 +119,7 @@ extern void (*Update_vol_only_sound)(void);
 int Pokey_sound_init(uint32 freq17, uint16 playback_freq, uint8 num_pokeys,
                      unsigned int flags
                      );
-void Pokey_process(void *sndbuffer, unsigned int sndn);
+void Pokey_process(void *sndbuffer, uint16 sndn);
 int Pokey_DoInit(void);
 void Pokey_set_mzquality(int quality);
